% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{calculate_frip}
\alias{calculate_frip}
\title{Calculate Fraction of Reads in Peaks}
\usage{
calculate_frip(reads_in_peaks, mapped_reads)
}
\arguments{
\item{reads_in_peaks}{Number of reads overlapping peaks.}

\item{mapped_reads}{Total number of mapped reads.}
}
\value{
A numeric FRiP value.
}
\description{
Calculates the fraction of mapped reads
that overlap called peaks.
}
